+++ src/commands/theme/component/generate.js
@@ -12,7 +12,7 @@ import Session from '../../../models/static/Session.js'
 import { copyFolder, getFolderFilesRecursively } from '../../../utils/FileUtils.js'
 import { getCLIRootFolderName, getPackageManifest, getPackageName, getPackageScope } from '../../../utils/NodeUtils.js'
 import { getValuesFromArgvOrToml } from '../../../utils/SessionUtils.js'
-import { info, logChildItem, logSeparator, logSpacer, logTitleItem } from '../../../utils/LoggerUtils.js'
+import { logChildItem, logSeparator, logSpacer, logTitleItem } from '../../../utils/LoggerUtils.js'
 
 export default class Generate extends BaseCommand {
   static description = 'Generate canvas files for new components'
@@ -86,15 +86,15 @@ export default class Generate extends BaseCommand {
       await copyFolder(sourcesPath, componentAbsolutePath, copyFolderOptions)
 
       logChildItem('complete')
-      info('\nThe following files were created:')
+      logSpacer()
+      logTitleItem('The following files were created:')
       const files = await getFolderFilesRecursively(componentPath)
-      files.forEach((file) => info(relative(componentPath, file)))
+      files.forEach((file) => logChildItem(relative(componentPath, file)))
       logSpacer()
 
-      info('Your new component is available at')
-      info('./' + componentPath)
+      logTitleItem('Your new component is available at')
+      logChildItem('./' + componentPath)
       logSeparator()
-      logSpacer()
     }
   }